+++ celescope/bulk_vdj/multi_bulk_vdj.py
@@ -44,7 +44,7 @@ class Multi_bulk_vdj(Multi):
          
     def consensus(self, sample):
         step = 'consensus'
-        fq = f'{self.outdir_dic[sample]["cutadapt"]}/{sample}_clean_2.fq{self.fq_suffix}'
+        fq = f'{self.outdir_dic[sample]["cutadapt"]}/{sample}_clean_2.fq'
         cmd_line = self.get_cmd_line(step, sample)
         cmd = (
             f'{cmd_line} '
@@ -71,7 +71,7 @@ class Multi_bulk_vdj(Multi):
         cmd_line = self.get_cmd_line(step, sample)
         productive_file = f'{self.outdir_dic[sample]["mapping_vdj"]}/{sample}_productive.tsv'
         airr_file = f'{self.outdir_dic[sample]["mapping_vdj"]}/{sample}_airr.tsv'
-        fq = f'{self.outdir_dic[sample]["barcode"]}/{sample}_2.fq{self.fq_suffix}'
+        fq = f'{self.outdir_dic[sample]["barcode"]}/{sample}_2.fq'
         cmd = (
             f'{cmd_line} '
             f'--productive_file {productive_file} '