+++ celescope/vdj/mkref.py
@@ -41,22 +41,15 @@ class Mkref:
         """
         Combine all V, all D and all J sequences, respectively, into separate files:
         """
-        imgt_files = os.listdir()
         if self.type == "TR":
-            imgt_files = [i for i in imgt_files if i[:2]=="TR"]
-            assert len(imgt_files) == 5
             cmd_V = f"cat TRAV.fasta TRBV.fasta > {self.outdir}/TRV.fasta"
             cmd_D = f"cat TRBD.fasta > {self.outdir}/TRD.fasta" 
             cmd_J = f"cat TRAJ.fasta TRBJ.fasta > {self.outdir}/TRJ.fasta"
         elif self.type == "TRGD":
-            imgt_files = [i for i in imgt_files if i[:2]=="TR"]
-            assert len(imgt_files) == 5
             cmd_V = f"cat TRDV.fasta TRGV.fasta > {self.outdir}/TRV.fasta"
             cmd_D = f"cat TRDD.fasta > {self.outdir}/TRD.fasta" 
             cmd_J = f"cat TRDJ.fasta TRGJ.fasta > {self.outdir}/TRJ.fasta"            
         else:
-            imgt_files = [i for i in imgt_files if i[:2]=="IG"]
-            assert len(imgt_files) == 7
             cmd_V = f"cat IGHV.fasta IGKV.fasta IGLV.fasta > {self.outdir}/IGV.fasta"
             cmd_D = f"cat IGHD.fasta > {self.outdir}/IGD.fasta" 
             cmd_J = f"cat IGHJ.fasta IGKJ.fasta IGLJ.fasta > {self.outdir}/IGJ.fasta"